+++ src/main/java/org/eclipse/tractusx/ssi/lib/proof/LinkedDataProofValidation.java
@@ -27,6 +27,7 @@ import lombok.SneakyThrows;
 import org.eclipse.tractusx.ssi.lib.did.resolver.DidResolver;
 import org.eclipse.tractusx.ssi.lib.exception.UnsupportedSignatureTypeException;
 import org.eclipse.tractusx.ssi.lib.model.verifiable.Verifiable;
+import org.eclipse.tractusx.ssi.lib.model.verifiable.Verifiable.VerifiableType;
 import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredential;
 import org.eclipse.tractusx.ssi.lib.model.verifiable.presentation.VerifiablePresentation;
 import org.eclipse.tractusx.ssi.lib.proof.hash.HashedLinkedData;
@@ -76,7 +77,9 @@ public class LinkedDataProofValidation {
    */
   @SneakyThrows
   public boolean verify(Verifiable verifiable) {
-
+    if (verifiable.getProof() == null) {
+      throw new UnsupportedSignatureTypeException("Proof can't be empty");
+    }
     var type = verifiable.getProof().getType();
     IVerifier verifier = null;
 
@@ -96,6 +99,44 @@ public class LinkedDataProofValidation {
     final TransformedLinkedData transformedData = transformer.transform(verifiable);
     final HashedLinkedData hashedData = hasher.hash(transformedData);
 
-    return jsonLdValidator.validate(verifiable) && verifier.verify(hashedData, verifiable);
+    return jsonLdValidator.validate(verifiable)
+        && verifier.verify(hashedData, verifiable)
+        && validateVerificationMethodOfVC(verifiable);
+  }
+
+  /**
+   * This method is to validate the Verification Method of VC
+   *
+   * @param verifiable
+   * @return
+   * @throws UnsupportedSignatureTypeException
+   */
+  @SneakyThrows
+  private Boolean validateVerificationMethodOfVC(Verifiable verifiable) {
+    // Verifiable Presentation doesn't have an Issuer
+    if (verifiable.getType() == VerifiableType.VP) {
+      return true;
+    }
+    final VerifiableCredential vc = new VerifiableCredential(verifiable);
+    final String issuer = vc.getIssuer().toString();
+    final String verficationMethod = getVerificationMethod(verifiable);
+    final String[] splitVerificationMethod = verficationMethod.split("#");
+    return splitVerificationMethod[0].equals(issuer);
+  }
+
+  /**
+   * This method is to get the Verification Method of VC
+   *
+   * @param verifiable
+   * @return
+   * @throws UnsupportedSignatureTypeException
+   */
+  @SneakyThrows
+  private String getVerificationMethod(Verifiable verifiable) {
+    try {
+      return (String) verifiable.getProof().get("verificationMethod");
+    } catch (Exception e) {
+      throw new UnsupportedSignatureTypeException("Signature type is not supported");
+    }
   }
 }